requirements.txt
aioredis==1.3.1
aniso8601==7.0.0
argon2-cffi==19.2.0
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
daphne==2.5.0
Django==2.2
django-cors-headers==3.2.0
django-filter==2.3.0
django-graphql-auth==0.3.11
django-graphql-jwt==0.3.0
django-model-utils==4.0.0
django-service-objects==0.7.0
django-taggit==1.2.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.11.0
graphene-django-extras==0.4.9
graphene-file-upload==1.2.2
graphene-subscriptions==1.0.2
graphql-core==2.2.1
graphql-relay==2.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
Pillow==6.2.1
promise==2.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.0
Unidecode==1.1.1
wrapt==1.11.2
zope.interface==5.1.0
